Fahrenheit spelling correction The name Fahrenheit was mis-spelt as Farenheit in several places; these have been corrected.
The incorrect spelling was also used in the ASCOM.Utilities.Units enum so this value has been deprecated and a new enum element introduced that has the correct spelling and the same enum value. Existing programs will continue to run without change, but Programs that use the incorrect spelling will now throw a compile time error until the deprecated value is updated.

This is essentially a non-breaking change for already compiled programs and a breaking change if programs are recompiled.